Support external debug symbols
If there are no debug symbols in the ELF, look in the normal places
on the system for available symbols.

We attempt to match gdb's documented logic:
https://sourceware.org/gdb/onlinedocs/gdb/Separate-Debug-Files.html

Handle c types better. Go pointer types include a name so we don't
have to recursively resolve the underlying types. It seems that is
not the case for GCC produced symbols. For those we'll recursively
resolve the underlying type.